Fixes code view
